// tickethunt-webapp/src/app/model/event.ts
export type EventData = {
  eventId?: String | null | undefined;
  email?: string | null;
  eventName?: String;
  date?: String;
  time?: String;
  venue?: String;
  organizerName?: String;
  totalSeats?: Number;
  eventType?: string;
  description?: string;
  price?: Number;
  rating?: string;
  image?: any,
};